requirements-dev.txt
-r requirements.txt
pytest>=6.0
pytest-django>=4.0
pytest-cov>=3.0.0
pre-commit>=2.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0
mypy>=0.900
twine>=4.0.0
build>=0.8.0
